Update borrado logica + logica de si ya han sido borrados (cliente y solicitud)
Se ha añadido la fecha en la que se realiza el borrado logico (mediante la query usando FECHA_BORRADO = SYSDATE) para cliente y request y se ha añadido el metodo en el servicio para verificar si ya habian sido borrados previamente.

+++ src/main/java/com/babelgroup/renting/services/impl/RentingRequestServiceImpl.java
@@ -108,7 +108,15 @@ public class RentingRequestServiceImpl implements RentingRequestService {
         return this.rentingRequestMapper.numberOfDeniedRequest(rentingRequestId) <= 0;
     }
 
+    private boolean isRequestAlreadyDeleted(long rentingRequestId) {
+        int deletionStatus = rentingRequestMapper.getDeletionStatus(rentingRequestId);
+        return  deletionStatus == 1;
+    }
+
     public boolean deleteRentingRequest(long rentingRequestId) {
+        if (isRequestAlreadyDeleted(rentingRequestId)) {
+            return false;
+        }
         if (canDeleteRequestWithDeniedStatus(rentingRequestId)) {
             this.rentingRequestMapper.deleteRentingRequest(rentingRequestId);
             return true;
